ai_companion_mvp/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
jinja2
